Solved ft_printf

+++ ft_printf/ft_printf.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: wookim <wookim@student.42seoul.kr>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/23 11:23:29 by wookim            #+#    #+#             */
+/*   Updated: 2022/03/23 11:23:33 by wookim           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
@@ -7,10 +19,10 @@
 int	ft_putstr(char *str);
 int	ft_putchar(char c);
 int	ft_printf(const char *format, ...);
-int	ft_putnbr(long long n);
+int	ft_putnbr(int n);
 int	ft_hexadecimal(unsigned int nbr);
 int	ft_hexadecimal_2(unsigned int nbr);
-int	ft_unsigned(long long nbr);
+int	ft_unsigned(int nbr);
 int	ft_pointer(void *ptr);
 int	ft_pointerhexadecimal(unsigned long nbr);
 #endif
